# contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: AccessCheck Lishui API
  version: 1.0.0
paths:
...
  /api/scans:
    post:
      responses:
        "202": { description: queued }
        "422": { description: invalid URL or options }
...
  /api/runs/{runId}/ai-review:
    get:
      parameters:
        - { name: runId, in: path, required: true, schema: { type: string } }
        - { name: pageId, in: query, schema: { type: string } }
      responses:
        "200": { description: AI overlay batch and dynamic statistics }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerConfigId]
              additionalProperties: false
              properties:
                providerConfigId: { type: string }
                pageId: { type: string }
      responses:
        "201": { description: idempotently created AI overlay batch }
...
  /api/ai/batches/{batchId}:
    get:
      parameters:
        - { name: batchId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: AI batch and dynamic item statistics }
    post:
      parameters:
        - { name: batchId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              additionalProperties: false
              properties:
                action: { type: string, enum: [pause, resume, retry] }
      responses:
        "200": { description: AI batch action applied }
...